# openapispec.yaml
openapi: 3.1.0
...
info:
  title: melodyMingle
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
...
  '/events/{eventId}':
    parameters:
      - schema:
          type: string
        name: eventId
        in: path
        required: true
    get:
      summary: Get a specific event by ID
      tags: []
      responses: {}
      operationId: get-events-eventId
      x-stoplight:
        id: bp42nd2yj0ere
    put:
      summary: Update a specific event
      tags: []
      responses: {}
      operationId: put-events-eventId
      x-stoplight:
        id: j4r9zes4jog65
    delete:
      summary: Delete an event by ID
      tags: []
      responses: {}
      operationId: delete-events-eventId
      x-stoplight:
        id: ta89ywb767htm
...
  /auth/register:
    post:
      summary: Register a new user
      tags: []
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                        minimum:
                          type: integer
                        type:
                          type: string
                        inclusive:
                          type: boolean
                        exact:
                          type: boolean
                        message:
                          type: string
                        path:
                          type: array
                          items:
                            type: string
                x-examples:
                  Example 1:
                    errors:
                      - code: too_small
                        minimum: 6
                        type: string
                        inclusive: true
                        exact: false
                        message: Password must be at least 6 characters long
                        path:
                          - password
        2XX:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  x-stoplight:
                    id: 8f5r972yzgyge
                  type: object
                  properties:
                    id:
                      type: string
                      x-stoplight:
                        id: k3igsqtw5wa7p
                    eMail:
                      type: string
                      x-stoplight:
                        id: f8b4co8nbuonr
                      format: email
                    username:
                      type: string
                      x-stoplight:
                        id: 23mk59ds24cti
                    dateOfBirth:
                      type: string
                      x-stoplight:
                        id: k0mo6s0xnmcvh
                      format: date
                    passwordHash:
                      type: string
                      x-stoplight:
                        id: l4ow3wutk557e
              examples:
                Example 1:
                  value:
                    - id: 260ccc4e-0f26-4b78-ab1e-163ceb11f855
                      eMail: a@de.de
                      username: asdvb
                      dateOfBirth: '2321-02-23'
                      passwordHash: $2b$10$7Y2pZewqTLfyK5dgHual/uVxUmxbJyStO6J5WE5RWtCLJIiNL4cEe
      operationId: post-auth-register
      x-stoplight:
        id: gu0zctw5dllph
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                eMail:
                  type: string
                  x-stoplight:
                    id: g7mss49rvsmvb
                  format: email
                username:
                  type: string
                  x-stoplight:
                    id: p2xn22d0glxpg
                dateOfBirth:
                  type: string
                  x-stoplight:
                    id: 168bxjb0x714x
                  format: date
                password:
                  type: string
                  x-stoplight:
                    id: 0gfvnbol0abpa
            examples:
              Example 1:
                value:
                  eMail: user@example.com
                  username: string
                  dateOfBirth: '2019-08-24'
                  password: string
      parameters: []
  /auth/login:
    post:
      summary: Login a user with credentials
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    x-stoplight:
                      id: olf963wo5t5mb
      operationId: post-auth-login
      x-stoplight:
        id: fogtttxkbowgv
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                eMail:
                  type: string
                  x-stoplight:
                    id: h39v8jhp8xgnu
                  format: email
                password:
                  type: string
                  x-stoplight:
                    id: 6v2st6bvn48gg
...
  /auth/spotify/callback:
    get:
      summary: Link Spotify Account to User
      tags: []
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    x-stoplight:
                      id: dq7l7yp7hcspt
              examples:
                Example 1:
                  value:
                    message: code not provided!
      operationId: get-auth-spotify-callback
      x-stoplight:
        id: l9kxlng5kwry3
    parameters:
      - schema:
          type: string
        in: query
        name: userId
        description: id of the user to be authenticated
  /auth/spotify/redirect:
    get:
      summary: Redirect to Frontend to get User Info
      tags: []
      responses: {}
      operationId: get-auth-spotify-redirect
      x-stoplight:
        id: 0j8naj57icw9k
      parameters:
        - schema:
            type: string
          in: query
      description: Forwards query parameters to frontend redirect url endpoint
...